Invoke shell script using `system()` not `exec()`
Using `exec()` causes the current process to be replaced,
which causes the caller of ship-it to die unexpectedly.

This was never an issue before because our clients always loaded
ship-it as the last thing they did.  However, sapi recently
added an `ensure` block after loading ship-it, and that `ensure`
block silently failed to run, due to the exec.